cli: move cwl_runner to indenpendent module
* Since `cwl_runner` does not belong to the standard command list for reana
  it is moved to an independent module.
